feat (ResultView/showWinningNumber) : 당첨 통계를 표시하는 콘솔 기능
